Increase maintainability by reducing complexity in functions, in addition to breaking python_functions into update_ui. This doesn't affect any of the user interactions.
