+++ blendernc/nodes/math/BlenderNC_NT_math.py
@@ -5,12 +5,9 @@ from collections import defaultdict
 import bpy
 import numpy as np
 
+from blendernc.core.update_ui import update_node_tree, update_value_and_node_tree
 from blendernc.decorators import NodesDecorators
-from blendernc.python_functions import (
-    refresh_cache,
-    update_node_tree,
-    update_value_and_node_tree,
-)
+from blendernc.python_functions import refresh_cache
 
 operation_items = [
     ("Multiply", "Multiply", "", 1),
