+++ blendernc/nodes/selecting/BlenderNC_NT_sort.py
@@ -4,9 +4,9 @@ from collections import defaultdict
 
 import bpy
 
-from blendernc.decorators import NodesDecorators
+from blendernc.core.update_ui import update_node_tree, update_value_and_node_tree
+from blendernc.decorators import DrawDecorators, NodesDecorators
 from blendernc.get_utils import get_possible_dims
-from blendernc.python_functions import update_node_tree, update_value_and_node_tree
 
 
 class BlenderNC_NT_sort(bpy.types.Node):
@@ -51,15 +51,9 @@ class BlenderNC_NT_sort(bpy.types.Node):
         print("Removing node ", self, ", Goodbye!")
 
     # Additional buttons displayed on the node.
+    @DrawDecorators.is_connected
     def draw_buttons(self, context, layout):
-        if (
-            self.inputs[0].is_linked
-            and self.inputs[0].links
-            and self.blendernc_dataset_identifier
-        ):
-            blendernc_dict = self.inputs[0].links[0].from_node.blendernc_dict
-            if blendernc_dict:
-                layout.prop(self, "blendernc_dims", text="")
+        layout.prop(self, "blendernc_dims", text="")
 
     # Detail buttons in the sidebar.
     # If this function is not defined,
